+++ src/Narochno.Primitives.Parsing/IParserLibrary.cs
@@ -1,0 +1,9 @@
+﻿using Narochno.Primitives.Parsing.Parsers;
+
+namespace Narochno.Primitives.Parsing
+{
+    public interface IParserLibrary
+    {
+        IParser FindParser<TType>();
+    }
+}
