+++ src/Narochno.Primitives.Parsing/Parsers/Parser.cs
@@ -1,4 +1,4 @@
-﻿namespace Narochno.Primitives.Parsing
+﻿namespace Narochno.Primitives.Parsing.Parsers
 {
     /// <summary>
     /// The base class for all <typeparamref name="TType"/> parsers
@@ -11,7 +11,7 @@
 
         /// <summary>
         /// Convert from <paramref name="input"/> to the type <typeparamref name="TType"/>
-        /// Should throw <see cref="System.ArgumentException"/> if invalid.
+        /// Should throw if invalid.
         /// </summary>
         /// <param name="input">The input string to convert from</param>
         /// <returns>A converted <typeparamref name="TType"/> object</returns>
@@ -19,7 +19,7 @@
 
         /// <summary>
         /// Convert from <paramref name="input"/> to the type <typeparamref name="TType"/>
-        /// Should not throw on exception, instead return an unset <see cref="Optional{TType}" />.
+        /// Should not throw, instead return an unset <see cref="Optional{TType}" />.
         /// </summary>
         /// <param name="input">The input string to convert from</param>
         /// <returns>An <see cref="Optional{TType}" /></returns>
